pyproject.toml
[project]
name = "text-embedding-visualization-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.6",
    "pandas>=2.2.3",
    "pre-commit>=4.2.0",
    "pydantic>=2.11.3",
    "sentence-transformers>=4.1.0",
    "torch>=2.7.0",
    "tqdm>=4.67.1",
    "streamlit>=1.44.1",
    "umap-learn>=0.5.7",
    "pacmap>=0.8.0",
    "scikit-learn>=1.6.1",
    "trimap>=1.1.4",
    "plotly>=6.0.1",
]




[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "notebook>=7.4.1",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-dotenv>=0.5.2",
    "ruff>=0.11.6",
]


[tool.ruff]
line-length=120
